LocalAI_.server/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "localai-server"
version = "0.1.0"
description = "Multimodal inference API skeleton built with FastAPI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic>=2",
    "httpx",
    "python-dotenv",
]

[project.optional-dependencies]
dev = []
